Refactored the main code
Split app_main into a new module.
Improved and cleaned up
