+++ src/amulet_editor/__main__.py
@@ -1,31 +1,42 @@
 #!/usr/bin/env python3
 
-
 def _on_error(e):
     """Code to handle errors"""
-    err_list = []
     try:
         import traceback
+        import sys
 
-        err_list.append(traceback.format_exc())
-    except ImportError:
-        pass
-    if isinstance(e, ImportError):
-        err_list.append(
-            f"Failed to import requirements. Check that you extracted correctly."
+    except ImportError as e:
+        # Something has gone seriously wrong
+        print(e)
+        print("Failed to import requirements. Check that you extracted correctly.")
+        input("Press ENTER to continue.")
+    else:
+        err = "\n".join(
+            [traceback.format_exc()]
+            + ["Failed to import requirements. Check that you extracted correctly."]
+            * isinstance(e, ImportError)
+            + [str(e)]
         )
-    err_list.append(str(e))
-    err = "\n".join(err_list)
-    print(err)
-    # TODO: fix this path to a writable location
-    with open("crash.log", "w") as f:
-        f.write(err)
-    input("Press ENTER to continue.")
-    sys.exit(1)
+        print(err)
+        try:
+            with open("crash.log", "w") as f:
+                f.write(err)
+        except OSError:
+            pass
+        input("Press ENTER to continue.")
+        sys.exit(1)
 
 
 try:
+    from multiprocessing import freeze_support
+
+    freeze_support()
     import sys
+    import logging
+    import faulthandler
+
+    faulthandler.enable()
 
     if sys.version_info[:2] < (3, 9):
         raise Exception("Must be using Python 3.9+")
@@ -35,19 +46,12 @@ except Exception as e_:
 
 def main():
     try:
-        from multiprocessing import freeze_support
-
-        freeze_support()
-        import logging
-        import faulthandler
-        faulthandler.enable()
-
-        from amulet_editor.application._app import app_main
-        from amulet_editor.models.widgets import AmuletTracebackDialog
-
         # Initialise logging at the highest level until configured otherwise.
         logging.basicConfig(level=logging.WARNING, format="%(levelname)s - %(message)s")
         logging.getLogger().setLevel(logging.WARNING)
+
+        from amulet_editor.application._main import app_main
+        from amulet_editor.models.widgets import AmuletTracebackDialog
 
     except Exception as e:
         _on_error(e)
